requirements.txt
tensorflow-gpu
ray
jupyter
requests
pillow
psutil
setproctitle
torch
torchvision
cython
